chamorajg/requirements.txt
torch==1.9.0
torchvision==0.10.0
pandas==1.1.3
timm==0.4.12
opencv_python_headless==4.3.0.36
tqdm==4.62.0
wandb==0.11.1
numpy==1.20.3
apex==0.9.10dev
Pillow==9.0.0
PyYAML==6.0
